added vertex input added object uniform

+++ vm/native/primitive/vec2.cpp
@@ -13,11 +13,11 @@ const vec2 vec2::ONE(0, 0);
 const vec2 vec2::EX(1, 0);
 const vec2 vec2::EY(0, 1);
 
-void vec2::set_x(double _x) { x = _x; }
-double vec2::get_x() const { return x; }
+void vec2::set_x(float _x) { x = _x; }
+float vec2::get_x() const { return x; }
 
-void vec2::set_y(double _y) { y = _y; }
-double vec2::get_y() const { return y; }
+void vec2::set_y(float _y) { y = _y; }
+float vec2::get_y() const { return y; }
 
 void vec2::set_xx(vec2 v) { x = v.x; x = v.y; }
 void vec2::set_yx(vec2 v) { y = v.x; x = v.y; }
@@ -27,12 +27,12 @@ vec2 vec2::get_xx() const { return vec2(x, x); }
 vec2 vec2::get_yx() const { return vec2(y, x); }
 vec2 vec2::get_yy() const { return vec2(y, y); }
 
-double vec2::dot(vec2 v) const
+float vec2::dot(vec2 v) const
 {
     return x * v.x + y * v.y;
 }
 
-double vec2::length() const
+float vec2::length() const
 {
     return std::sqrt(length_squared());
 }
@@ -41,15 +41,15 @@ void vec2::register_to_db(vm::ClassDB& db) {
         const int16_t id = REGISTER_CLASS(vec2);
 		
 		REGISTER_CONSTRUCTOR(id, vec2());
-		REGISTER_CONSTRUCTOR(id, vec2(double x, double y));
+		REGISTER_CONSTRUCTOR(id, vec2(float x, float y));
 
-		REGISTER_VARIABLE(id, double, x);
-		REGISTER_VARIABLE(id, double, y);
+		REGISTER_VARIABLE(id, float, x);
+		REGISTER_VARIABLE(id, float, y);
 
-        REGISTER_METHOD(id, vec2, set_x, void (vec2::*)(double x));
-        REGISTER_METHOD(id, vec2, get_x, double (vec2::*)() const);
-        REGISTER_METHOD(id, vec2, set_y, void (vec2::*)(double y));
-        REGISTER_METHOD(id, vec2, get_y, double (vec2::*)() const);
+        REGISTER_METHOD(id, vec2, set_x, void (vec2::*)(float x));
+        REGISTER_METHOD(id, vec2, get_x, float (vec2::*)() const);
+        REGISTER_METHOD(id, vec2, set_y, void (vec2::*)(float y));
+        REGISTER_METHOD(id, vec2, get_y, float (vec2::*)() const);
 
         REGISTER_METHOD(id, vec2, set_xx, void (vec2::*)(vec2 xx));
         REGISTER_METHOD(id, vec2, set_yx, void (vec2::*)(vec2 yx));
@@ -60,8 +60,8 @@ void vec2::register_to_db(vm::ClassDB& db) {
         REGISTER_METHOD(id, vec2, get_yy, vec2 (vec2::*)() const);
 
         REGISTER_METHOD(id, vec2, dot);
-        REGISTER_METHOD(id, vec2, length_squared, double (vec2::*)() const);
-        REGISTER_METHOD(id, vec2, length, double (vec2::*)() const);
+        REGISTER_METHOD(id, vec2, length_squared, float (vec2::*)() const);
+        REGISTER_METHOD(id, vec2, length, float (vec2::*)() const);
 
         REGISTER_METHOD(id, vec2, operator+, vec2 (vec2::*)(const vec2& other) const);
         REGISTER_METHOD(id, vec2, operator-, vec2 (vec2::*)() const);
